refactor: split resample into multiple modules
